CSS refactor content encapsulation
- Made the outermost div only have {className ?? ""} for PageSection and SectionArticle.
- Moved outermost (h-screen flex > Navbar + flex grow) into _app solely rather than growing ProtectedRoute to avoid confusing behaviour when ProtectedRoute is not needed.
- Edited the home page grid layout to avoid nesting grids (grid-gap is better than margin)
- Accomodated encapsulation changes for elements using PageSection
