+++ web/pages/_app.tsx
@@ -18,7 +18,9 @@ const App = ({ Component, pageProps }: AppProps) => {
             <QueryClientProvider client={queryClient}>
                 <div className="flex flex-col h-screen bg-gray-100">
                     <Navbar />
-                    <Component {...pageProps} />
+                    <div className="grow overflow-auto">
+                        <Component {...pageProps} />
+                    </div>
                 </div>
             </QueryClientProvider>
         </AuthProvider>
