+++ web/pages/home.tsx
@@ -14,19 +14,15 @@ const Home = () => {
 
     return (
         <ProtectedRoute>
-            <div className="w-full h-full flex grow flex-col p-4 space-y-4">
+            <div className="h-full p-4 flex flex-col space-y-4">
                 <Header>
                     Welcome back {auth.user?.profile.given_name ?? "Unknown"}!
                 </Header>
-                <div className="grid grid-cols-7 h-full w-full space-x-4">
+                <div className="grow grid grid-cols-7 grid-rows-2 gap-4">
                     <Recents />
-                    <div className="grid grid-rows-2 col-span-4 space-y-4">
-                        <div className="grid grid-cols-2 span-4 space-x-4">
-                            <SpacedRepetition />
-                            <Groups />
-                        </div>
-                        <Calendar />
-                    </div>
+                    <SpacedRepetition />
+                    <Groups />
+                    <Calendar />
                 </div>
             </div>
         </ProtectedRoute>
